requirements.txt
requests>=2.27.1
pytest>=7.0.1
jsonschema>=4.4.0
allure-pytest==2.9.45